:recycle: rename to components/atoms
