+++ lib/calendar/calendar.dart
@@ -10,9 +10,9 @@ import 'package:Pilll/entity/weekday.dart';
 import 'package:Pilll/service/diary.dart';
 import 'package:Pilll/store/confirm_diary_sheet.dart';
 import 'package:Pilll/store/diaries.dart';
-import 'package:Pilll/theme/color.dart';
-import 'package:Pilll/theme/font.dart';
-import 'package:Pilll/theme/text_color.dart';
+import 'package:Pilll/components/atoms/color.dart';
+import 'package:Pilll/components/atoms/font.dart';
+import 'package:Pilll/components/atoms/text_color.dart';
 import 'package:Pilll/util/today.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
