+++ lib/components/atoms/button.dart
@@ -1,5 +1,5 @@
-import 'package:Pilll/theme/color.dart';
-import 'package:Pilll/theme/font.dart';
+import 'package:Pilll/components/atoms/color.dart';
+import 'package:Pilll/components/atoms/font.dart';
 import 'package:flutter/material.dart';
 
 abstract class ButtonTextStyle {
