+++ lib/components/pill_sheet_type_select_page.dart
@@ -1,8 +1,8 @@
 import 'package:Pilll/initial_setting/pill_sheet.dart';
 import 'package:Pilll/entity/pill_sheet_type.dart';
-import 'package:Pilll/theme/color.dart';
-import 'package:Pilll/theme/font.dart';
-import 'package:Pilll/theme/text_color.dart';
+import 'package:Pilll/components/atoms/color.dart';
+import 'package:Pilll/components/atoms/font.dart';
+import 'package:Pilll/components/atoms/text_color.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
