+++ lib/diary/post_diary_page.dart
@@ -4,9 +4,9 @@ import 'package:Pilll/state/diary.dart';
 import 'package:Pilll/store/diaries.dart';
 import 'package:Pilll/store/post_diary.dart';
 import 'package:Pilll/style/button.dart';
-import 'package:Pilll/theme/color.dart';
-import 'package:Pilll/theme/font.dart';
-import 'package:Pilll/theme/text_color.dart';
+import 'package:Pilll/components/atoms/color.dart';
+import 'package:Pilll/components/atoms/font.dart';
+import 'package:Pilll/components/atoms/text_color.dart';
 import 'package:Pilll/util/formatter/date_time_formatter.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
