+++ lib/home/home.dart
@@ -1,8 +1,8 @@
 import 'package:Pilll/calendar/calendar_page.dart';
 import 'package:Pilll/record/record_page.dart';
 import 'package:Pilll/settings/settings.dart';
-import 'package:Pilll/theme/color.dart';
-import 'package:Pilll/theme/text_color.dart';
+import 'package:Pilll/components/atoms/color.dart';
+import 'package:Pilll/components/atoms/text_color.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
