+++ lib/initial_setting/initial_setting.dart
@@ -1,8 +1,8 @@
 import 'package:Pilll/style/button.dart';
-import 'package:Pilll/theme/color.dart';
+import 'package:Pilll/components/atoms/color.dart';
 import 'package:Pilll/initial_setting/initial_setting_1.dart';
-import 'package:Pilll/theme/font.dart';
-import 'package:Pilll/theme/text_color.dart';
+import 'package:Pilll/components/atoms/font.dart';
+import 'package:Pilll/components/atoms/text_color.dart';
 import 'package:flutter/material.dart';
 
 class InitialSetting extends StatelessWidget {
