+++ lib/initial_setting/initial_setting_2.dart
@@ -1,11 +1,11 @@
 import 'package:Pilll/application/router.dart';
 import 'package:Pilll/store/initial_setting.dart';
 import 'package:Pilll/style/button.dart';
-import 'package:Pilll/theme/color.dart';
+import 'package:Pilll/components/atoms/color.dart';
 import 'package:Pilll/initial_setting/initial_setting_3.dart';
 import 'package:Pilll/components/pill/pill_sheet.dart';
-import 'package:Pilll/theme/font.dart';
-import 'package:Pilll/theme/text_color.dart';
+import 'package:Pilll/components/atoms/font.dart';
+import 'package:Pilll/components/atoms/text_color.dart';
 import 'package:Pilll/util/today.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
