+++ lib/initial_setting/initial_setting_4.dart
@@ -2,9 +2,9 @@ import 'package:Pilll/application/router.dart';
 import 'package:Pilll/entity/initial_setting.dart';
 import 'package:Pilll/store/initial_setting.dart';
 import 'package:Pilll/style/button.dart';
-import 'package:Pilll/theme/color.dart';
-import 'package:Pilll/theme/font.dart';
-import 'package:Pilll/theme/text_color.dart';
+import 'package:Pilll/components/atoms/color.dart';
+import 'package:Pilll/components/atoms/font.dart';
+import 'package:Pilll/components/atoms/text_color.dart';
 import 'package:Pilll/util/formatter/date_time_formatter.dart';
 import 'package:Pilll/util/toolbar/date_time_picker.dart';
 import 'package:flutter/cupertino.dart';
