+++ lib/initial_setting/pill_sheet.dart
@@ -1,7 +1,7 @@
-import 'package:Pilll/theme/color.dart';
+import 'package:Pilll/components/atoms/color.dart';
 import 'package:Pilll/entity/pill_sheet_type.dart';
-import 'package:Pilll/theme/font.dart';
-import 'package:Pilll/theme/text_color.dart';
+import 'package:Pilll/components/atoms/font.dart';
+import 'package:Pilll/components/atoms/text_color.dart';
 import 'package:flutter/material.dart';
 
 class PillSheet extends StatelessWidget {
