+++ lib/store/confirm_diary_sheet.dart
@@ -5,9 +5,9 @@ import 'package:Pilll/state/diary.dart';
 import 'package:Pilll/store/diaries.dart';
 import 'package:Pilll/store/post_diary.dart';
 import 'package:Pilll/style/button.dart';
-import 'package:Pilll/theme/color.dart';
-import 'package:Pilll/theme/font.dart';
-import 'package:Pilll/theme/text_color.dart';
+import 'package:Pilll/components/atoms/color.dart';
+import 'package:Pilll/components/atoms/font.dart';
+import 'package:Pilll/components/atoms/text_color.dart';
 import 'package:Pilll/util/formatter/date_time_formatter.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
